Moved ComplementaryFormsAggregationHelper to own file
